/incidents returns minimal representation. Split into files.
